Handle properly precision 0 for GeoTileGrid
