+++ x-pack/plugin/spatial/src/main/java/org/elasticsearch/xpack/spatial/search/aggregations/bucket/geogrid/AbstractGeoHashGridTiler.java
@@ -23,9 +23,6 @@ abstract class AbstractGeoHashGridTiler extends GeoGridTiler {
     /** check if the provided hash is in the solution space of this tiler */
     protected abstract boolean validHash(String hash);
 
-    /** Max size of the solution space */
-    protected abstract long getMaxHashes();
-
     @Override
     public long encode(double x, double y) {
         return Geohash.longEncode(x, y, precision);
@@ -34,17 +31,16 @@ abstract class AbstractGeoHashGridTiler extends GeoGridTiler {
     @Override
     public int setValues(GeoShapeCellValues values, GeoShapeValues.GeoShapeValue geoValue) {
 
+        if (precision == 0) {
+          return 1;
+        }
         GeoShapeValues.BoundingBox bounds = geoValue.boundingBox();
         assert bounds.minX() <= bounds.maxX();
 
-        // TODO: optimize for when a whole shape (not just point) fits in a single tile an
+        // TODO: optimize for when a  shape fits in a single tile an
         //  for when brute-force is expected to be faster than rasterization, which
         //  is when the number of tiles expected is less than the precision
 
-        // optimization for setting just one value for when the shape represents a point
-        if (bounds.minX() == bounds.maxX() && bounds.minY() == bounds.maxY()) {
-            return setValue(values, geoValue, bounds);
-        }
         return setValuesByRasterization("", values, 0, geoValue);
     }
 
@@ -86,7 +82,7 @@ abstract class AbstractGeoHashGridTiler extends GeoGridTiler {
         return 0;
     }
 
-    protected GeoRelation relateTile(GeoShapeValues.GeoShapeValue geoValue, String hash) {
+    private GeoRelation relateTile(GeoShapeValues.GeoShapeValue geoValue, String hash) {
         return validHash(hash) ? geoValue.relate(Geohash.toBoundingBox(hash)) : GeoRelation.QUERY_DISJOINT;
     }
 
@@ -126,7 +122,7 @@ abstract class AbstractGeoHashGridTiler extends GeoGridTiler {
     }
 
     private int getNumTilesAtPrecision(int finalPrecision, int currentPrecision) {
-        final long numTilesAtPrecision  = Math.min((long) Math.pow(32, finalPrecision - currentPrecision) + 1, getMaxHashes());
+        final long numTilesAtPrecision  = Math.min((long) Math.pow(32, finalPrecision - currentPrecision) + 1, getMaxCells());
         if (numTilesAtPrecision > Integer.MAX_VALUE) {
             throw new IllegalArgumentException("Tile aggregation array overflow");
         }
