+++ x-pack/plugin/spatial/src/main/java/org/elasticsearch/xpack/spatial/search/aggregations/bucket/geogrid/AbstractGeoTileGridTiler.java
@@ -26,9 +26,6 @@ abstract class AbstractGeoTileGridTiler extends GeoGridTiler {
     /** check if the provided tile is in the solution space of this tiler */
     protected abstract boolean validTile(int x, int y, int z);
 
-    /** Max size of the solution space */
-    protected abstract long getMaxTiles();
-
     @Override
     public long encode(double x, double y) {
         return GeoTileUtils.longEncode(x, y, precision);
@@ -57,6 +54,10 @@ abstract class AbstractGeoTileGridTiler extends GeoGridTiler {
             return 0;
         }
 
+        if (precision == 0) {
+            return validTile(0, 0, 0) ? 1 : 0;
+        }
+
         final int minXTile = GeoTileUtils.getXTile(bounds.minX(), tiles);
         final int minYTile = GeoTileUtils.getYTile(bounds.maxY(), tiles);
         final int maxXTile = GeoTileUtils.getXTile(bounds.maxX(), tiles);
@@ -71,7 +72,7 @@ abstract class AbstractGeoTileGridTiler extends GeoGridTiler {
         }
     }
 
-    protected GeoRelation relateTile(GeoShapeValues.GeoShapeValue geoValue, int xTile, int yTile, int precision) {
+    private GeoRelation relateTile(GeoShapeValues.GeoShapeValue geoValue, int xTile, int yTile, int precision) {
         return validTile(xTile, yTile, precision) ?
             geoValue.relate(GeoTileUtils.toBoundingBox(xTile, yTile, precision)) : GeoRelation.QUERY_DISJOINT;
     }
@@ -149,7 +150,7 @@ abstract class AbstractGeoTileGridTiler extends GeoGridTiler {
     }
 
     private int getNumTilesAtPrecision(int finalPrecision, int currentPrecision) {
-        final long numTilesAtPrecision  = Math.min(1L << (2 * (finalPrecision - currentPrecision)), getMaxTiles());
+        final long numTilesAtPrecision  = Math.min(1L << (2 * (finalPrecision - currentPrecision)), getMaxCells());
         if (numTilesAtPrecision > Integer.MAX_VALUE) {
             throw new IllegalArgumentException("Tile aggregation array overflow");
         }
